+++ src/test/java/org/opentripplanner/transit/raptor/_data/transit/TestTransitData.java
@@ -7,6 +7,9 @@ import java.util.HashSet;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Set;
+import java.util.function.IntFunction;
+import java.util.stream.Collectors;
+import javax.annotation.Nullable;
 import lombok.val;
 import org.opentripplanner.model.transfer.ConstrainedTransfer;
 import org.opentripplanner.model.transfer.TransferConstraint;
@@ -15,16 +18,19 @@ import org.opentripplanner.routing.algorithm.raptor.transit.cost.DefaultCostCalc
 import org.opentripplanner.routing.algorithm.raptor.transit.cost.McCostParamsBuilder;
 import org.opentripplanner.routing.algorithm.transferoptimization.model.TripStopTime;
 import org.opentripplanner.routing.algorithm.transferoptimization.services.TransferServiceAdaptor;
+import org.opentripplanner.transit.raptor._data.RaptorTestConstants;
 import org.opentripplanner.transit.raptor._data.debug.TestDebugLogger;
 import org.opentripplanner.transit.raptor.api.request.RaptorRequestBuilder;
 import org.opentripplanner.transit.raptor.api.transit.CostCalculator;
 import org.opentripplanner.transit.raptor.api.transit.IntIterator;
+import org.opentripplanner.transit.raptor.api.transit.RaptorConstrainedTransfer;
+import org.opentripplanner.transit.raptor.api.transit.RaptorPathConstrainedTransferSearch;
 import org.opentripplanner.transit.raptor.api.transit.RaptorRoute;
 import org.opentripplanner.transit.raptor.api.transit.RaptorTransfer;
 import org.opentripplanner.transit.raptor.api.transit.RaptorTransitDataProvider;
 
 @SuppressWarnings("UnusedReturnValue")
-public class TestTransitData implements RaptorTransitDataProvider<TestTripSchedule> {
+public class TestTransitData implements RaptorTransitDataProvider<TestTripSchedule>, RaptorTestConstants {
 
   private static final TransferConstraint GUARANTEED = new TransferConstraint(
           TransferPriority.ALLOWED, false, true, MAX_WAIT_TIME_NOT_SET
@@ -62,6 +68,38 @@ public class TestTransitData implements RaptorTransitDataProvider<TestTripSchedu
             costParamsBuilder.build(),
             stopBoarAlightCost()
     );
+  }
+
+  @Override
+  public RaptorPathConstrainedTransferSearch<TestTripSchedule> transferConstraintsSearch() {
+    return new RaptorPathConstrainedTransferSearch<>() {
+      @Nullable
+      @Override
+      public RaptorConstrainedTransfer findConstrainedTransfer(
+              TestTripSchedule fromTrip,
+              int fromStopPosition,
+              TestTripSchedule toTrip,
+              int toStopPosition
+      ) {
+        var list = routes.stream()
+                .flatMap(r -> r.listTransferConstraintsForwardSearch().stream())
+                .filter(tx -> tx.getSourceTrip().equals(fromTrip))
+                .filter(tx -> tx.getSourceStopPos() == fromStopPosition)
+                .filter(tx -> tx.getTrip().equals(toTrip))
+                .filter(tx -> tx.getStopPositionInPattern() == toStopPosition)
+                .collect(Collectors.toList());
+
+        if(list.isEmpty()) { return null; }
+        if(list.size() == 1) { return list.get(0); }
+        throw new IllegalStateException("More than on transfers found: " + list);
+      }
+    };
+  }
+
+  @Override
+  public IntFunction<String> stopIndexTranslatorForDebugging() {
+    // Index is translated: 1->'A', 2->'B', 3->'C' ...
+    return this::stopIndexToName;
   }
 
   public TestRoute getRoute(int index) {